+++ msi.gama.jogl/src/msi/gama/jogl/utils/GraphicDataType/MyImage.java
@@ -1,28 +1,24 @@
 package msi.gama.jogl.utils.GraphicDataType;
 
-
 import java.awt.image.BufferedImage;
-
 import msi.gama.metamodel.agent.IAgent;
 import msi.gama.metamodel.shape.GamaPoint;
 
-
-
-
-
 public class MyImage {
-	
+
 	public BufferedImage image;
 	public IAgent agent;
 	public int textureId;
 	public boolean isTextured;
 	public float x;
 	public float y;
-	public float z;
+	public double z;
 	public float alpha;
-	public float width;
-	public float height;
+	public double width;
+	public double height;
+	// public boolean isDynamic;
 	public Integer angle = 0;
 	public String name;
-	public GamaPoint offSet = new GamaPoint(0,0,0);
+	public GamaPoint offSet = new GamaPoint(0, 0, 0);
+	// public Texture texture;
 }
