+++ msi.gama.jogl/src/msi/gama/jogl/utils/GraphicDataType/MyString.java
@@ -1,12 +1,9 @@
 package msi.gama.jogl.utils.GraphicDataType;
 
-import java.awt.Color;
-
-
 public class MyString {
-	
+
 	public String string;
-	
-	public float x,y,z;
-	
+
+	public double x, y, z;
+
 }
